// src/shared/enums/reminderTime.enum.ts
export enum ReminderTimeEnum {
  never_remind = 'none',
  five_minute_before = '5',
  ten_minute_before = '10',
  fifteen_minute_before = '15',
  thirty_minute_before = '30',
  one_hour_before = '60',
  two_hour_before = '120',
  three_hour_before = '180',
  four_hour_before = '240',
  five_hour_before = '300',
  six_hour_before = '360',
  seven_hour_before = '420',
  eight_hour_before = '480',
  nine_hour_before = '540',
  ten_hour_before = '600',
  eleven_hour_before = '660',
  twelve_hour_before = '720',
  thirteen_hour_before = '780',
  fourteen_hour_before = '840',
  fifteen_hour_before = '900',
  sixteen_hour_before = '960',
  seventeen_hour_before = '1020',
  eighteen_hour_before = '1080',
  nineteen_hour_before = '1140',
  twenty_hour_before = '1200',
  twenty_one_hour_before = '1260',
  twenty_two_hour_before = '1320',
  twenty_three_hour_before = '1380',
  one_day_before = '1440',
  on_a_day_of_task = 'on_a_day_of_task',
  on_a_day_of_event = 'on_a_day_of_event',
  on_a_day_of_meeting = 'on_a_day_of_meeting',
  one_week_before = '10080',
  two_days_before = '2880',
  three_days_before = '4320',
  four_days_before = '5760',
  five_days_before = '7200',
}