// src/shared/enums/documentEntityType.enum.ts
export enum DocumentsEntityTypeEnum {
  event_document = 'event_document',
  task_document = 'task_document',
  payment_document = 'payment_document',
  todo_document = 'todo_document',
  comment_document = 'comment_document',
  feedback_document = 'feedback_document',
  friend_document = 'friend_document',
  insurance_document = 'insurance_document',
  insurance_card_front_document = 'insurance_card_front_document',
  insurance_card_back_document = 'insurance_card_back_document',
  license_document = 'license_document',
  sticker_document = 'sticker_document',
  transport_document = 'transport_document',
  message_document = 'message_document',
  user_document = 'user_document',
  section_document = 'section_document',
  document_body_arts = 'document_body_arts',
}